Removed windows warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1402 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/widget.cpp
@@ -96,7 +96,7 @@ void Widget::update(const float DELTA)
 {
     glPushMatrix();
 
-    glTranslatef ( m_x , m_y , 0);
+    glTranslatef ( (GLfloat)m_x , (GLfloat)m_y , 0);
 
     /*Handle delta time dependant features*/
     if(m_text_scale > MIN_TEXT_SCALE)
